fix: moved all get activites logic in separate file
